requirements.txt
adal==1.2.7
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiomultiprocess==0.9.1
aiosignal==1.3.2
annotated-types==0.7.0
applicationinsights==0.11.10
APScheduler==3.11.0
argcomplete==3.5.3
async-timeout==5.0.1
attrs==25.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==3.2.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.21.3
azure-graphrbac==0.61.2
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==1.0.0
azure-mgmt-appconfiguration==0.7.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-appplatform==8.0.0
azure-mgmt-authorization==1.0.0
azure-mgmt-automation==0.1.1
azure-mgmt-batch==15.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==11.0.0
azure-mgmt-compute==19.0.0
azure-mgmt-containerinstance==7.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==15.1.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==6.4.0
azure-mgmt-costmanagement==1.0.0
azure-mgmt-databricks==1.0.0b1
azure-mgmt-datafactory==1.1.0
azure-mgmt-datalake-analytics==0.5.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-desktopvirtualization==1.1.0
azure-mgmt-dns==8.0.0b1
azure-mgmt-eventgrid==10.3.0b4
azure-mgmt-eventhub==11.2.0
azure-mgmt-frontdoor==1.2.0
azure-mgmt-hdinsight==7.0.0
azure-mgmt-iothub==1.0.0
azure-mgmt-keyvault==8.0.0
azure-mgmt-kusto==2.2.0
azure-mgmt-logic==9.0.0
azure-mgmt-machinelearningservices==2.0.0b2
azure-mgmt-managementgroups==1.0.0b1
azure-mgmt-monitor==2.0.0
azure-mgmt-msi==1.0.0
azure-mgmt-network==28.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==2.5.0
azure-mgmt-recoveryservicesbackup==7.0.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==12.0.0
azure-mgmt-resource==16.1.0
azure-mgmt-resourcegraph==7.0.0
azure-mgmt-search==8.0.0
azure-mgmt-security==1.0.0
azure-mgmt-servicebus==8.2.1
azure-mgmt-servicefabric==1.0.0
azure-mgmt-signalr==0.4.0
azure-mgmt-sql==1.0.0
azure-mgmt-storage==17.1.0
azure-mgmt-streamanalytics==1.0.0
azure-mgmt-subscription==1.0.0
azure-mgmt-synapse==2.0.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==3.0.0
azure-nspkg==3.0.2
azure-storage-blob==12.24.1
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-file-share==12.20.1
azure-storage-queue==12.12.0
bc-detect-secrets==1.5.36
bc-jsonpath-ng==1.6.1
bc-python-hcl2==0.4.2
beartype==0.19.0
beautifulsoup4==4.13.3
blinker==1.9.0
boolean.py==4.0
boto3==1.36.11
botocore==1.36.11
c7n==0.9.43
c7n_azure==0.7.42
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
checkov==3.2.372
click==8.1.8
click-option-group==0.5.6
cloudevents==1.11.0
cloudsplaining==0.7.0
colorama==0.4.6
ConfigArgParse==1.7
contextlib2==21.6.0
cryptography==43.0.3
cyclonedx-python-lib==7.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.9
docker==7.1.0
dockerfile-parse==2.0.1
docutils==0.18.1
dpath==2.1.3
Flask==3.1.0
frozenlist==1.5.0
functions-framework==3.8.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.4.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
junit-xml==1.9
lark==1.2.2
license-expression==30.4.1
Markdown==3.7
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.1.0
netaddr==0.7.20
networkx==2.6.3
numpy==2.2.3
oauthlib==3.2.2
openai==0.28.1
orjson==3.10.15
packageurl-python==0.13.4
packaging==23.2
ply==3.11
policy_sentry==0.13.2
portalocker==2.10.1
prettytable==3.14.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
py-serializable==1.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.5.0
pycep-parser==0.5.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.1
pyston==2.3.5
pyston-autoload==2.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
rustworkx==0.13.2
s3transfer==0.11.2
schema==0.7.5
semantic-version==2.10.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.6
spdx-tools==0.8.3
tabulate==0.9.0
termcolor==2.3.0
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
unidiff==0.7.5
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.20
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
xmltodict==0.14.2
yarl==1.18.3
zipp==3.20.2
